Fix raddius of drag item

+++ interfaces/innerkits/interaction/include/drag_data.h
@@ -135,7 +135,7 @@ struct PreviewStyle {
     std::vector<PreviewType> types;
     uint32_t foregroundColor { 0 };
     int32_t opacity { -1 };
-    int32_t radius { -1 };
+    float radius { -1 };
     float scale { -1 };
 
     bool operator == (const PreviewStyle &other) const
